pyproject.toml
[project]
name = "multiagententerprise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "fastmcp>=3.2.4",
    "grandalf>=0.8",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.9",
    "langsmith>=0.7.37",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
]
